feature: move is weekend and is holiday to utils
